+++ rules-tests/Order/StmtOrderTest.php
@@ -78,9 +78,9 @@ final class StmtOrderTest extends AbstractKernelTestCase
 
     /**
      * @dataProvider dataProvider
-     * @param array<int, string> $desiredStmtOrder
-     * @param array<int, string> $currentStmtOrder
-     * @param array<int, int> $expected
+     * @param string[] $desiredStmtOrder
+     * @param string[] $currentStmtOrder
+     * @param int[] $expected
      */
     public function testCreateOldToNewKeys(array $desiredStmtOrder, array $currentStmtOrder, array $expected): void
     {