+++ rules/DeadCode/PhpDoc/TagRemover/VarTagRemover.php
@@ -121,9 +121,10 @@ final class VarTagRemover
     {
         if ($varTagValueNode->type instanceof BracketsAwareUnionTypeNode) {
             foreach ($varTagValueNode->type->types as $type) {
-                if ($type instanceof SpacingAwareArrayTypeNode && $this->isArrayOfExistingClassNode($node, $type)) {
-                    return true;
+                if (! $type instanceof SpacingAwareArrayTypeNode) {
+                    return false;
                 }
+                return $this->isArrayOfExistingClassNode($node, $type);
             }
         }